💄 Criando tela de menu pratos
