+++ src/Pages/AdminLayout/index.tsx
@@ -3,10 +3,10 @@ import styled from "styled-components"
 import { useContext, useEffect, useState } from "react";
 import { ThemeContext } from "../../context/themeContext";
 import HeaderMain from "../../components/layout/Header";
-import { ModalContext } from "../../context/modalContext";
 import Sidebar from "../../components/layout/Sidebar";
 import { Outlet } from "react-router-dom";
 import { Content, MainContainer, MainSection } from "./AdminLayout.styles";
+import Modal from "../../components/common/Modal";
 
 
 
@@ -33,6 +33,7 @@ const Layout = () => {
         <Content>
           <Outlet />
         </Content>
+        <Modal />
       </MainSection>
     </MainContainer>
   )
