+++ src/components/common/Modal/index.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from 'react'
 import { ButtonAction, Centralize, ModalDialog, Overlay, Paragraph, SpacingContents } from './Modal.styles'
-import { ModalContext } from '../../../context/modalContext'
+import { ModalContext, ModalContextProps } from '../../../context/modalContext'
 
 const Modal = () => {
-    const modalContext = useContext(ModalContext)
+    const {isOpen, closeModal} = useContext(ModalContext) as ModalContextProps
 
-    if(!modalContext?.isOpen){
+    if(!isOpen){
         return null
     }
 
@@ -15,8 +15,8 @@ const Modal = () => {
                 <ModalDialog>
                     <Paragraph>Você tem certeza que deseja realizar esta ação?</Paragraph>
                     <SpacingContents>
-                        <ButtonAction onClick={modalContext.closeModal} style={{ backgroundColor: '#00D982' }} >Sim</ButtonAction>
-                        <ButtonAction onClick={modalContext.closeModal} style={{ backgroundColor: '#C90000' }}>Não</ButtonAction>
+                        <ButtonAction onClick={closeModal} style={{ backgroundColor: '#00D982' }}>Sim</ButtonAction>
+                        <ButtonAction onClick={closeModal} style={{ backgroundColor: '#C90000' }}>Não</ButtonAction>
                     </SpacingContents>
                 </ModalDialog>
             </Centralize>
